Make installable via astro add
