+++ inlang/source-code/paraglide/paraglide-js-adapter-astro/package.json
@@ -20,5 +20,8 @@
     },
     "peerDependencies": {
         "astro": "^4.0.0"
-    }
+    },
+    "keywords": [
+        "astro-integration"
+    ]
 }
