+++ inlang/source-code/paraglide/paraglide-js-adapter-astro/src/ambient.d.ts
@@ -11,6 +11,11 @@ declare module "paraglide-js-adapter-astro:runtime" {
 	export const sourceLanguageTag: string
 }
 
+declare module "astro:i18n" {
+	export const getLocaleByPath: (path: string) => string
+	export const getPathByLocale: (locale: string) => string
+}
+
 declare namespace App {
 	interface Locals {
 		paraglide: {
