--- inlang/source-code/paraglide/paraglide-js-adapter-astro/src/astro-i18n.d.ts
@@ -1,4 +1,0 @@
-declare module "astro:i18n" {
-	export const getLocaleByPath: (path: string) => string
-	export const getPathByLocale: (locale: string) => string
-}
